LayoutTests/media/modern-media-controls/icon-button/icon-button-focus-state.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing an <code>IconButton</code> has a blue background-color when focused.");

window.jsTestIsAsync = true;

const iconButton = new IconButton({ layoutDelegate: { layoutTraits: LayoutTraits.macOS } });
iconButton.iconName = Icons.Pause;
document.body.appendChild(iconButton.element);

debug("Confirm inital icon button color.");
shouldBeEqualToString("window.getComputedStyle(iconButton.element).backgroundColor", "rgb(192, 192, 192)");

iconButton.element.focus();

debug("Confirm icon button color changed after focus.");
shouldNotBeEqualToString("window.getComputedStyle(iconButton.element).backgroundColor", "rgb(192, 192, 192)");

finishJSTest();
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
